Move code from sphinx_copybutton.py to __init__.py
